Update rapidjson, and move code upwards the upstream's trunk.

+++ Source/Lutefisk3D/UI/FontFace.cpp
@@ -66,7 +66,7 @@ const FontGlyph* FontFace::GetGlyph(unsigned c)
         return nullptr;
 }
 
-short FontFace::GetKerning(unsigned c, unsigned d) const
+float FontFace::GetKerning(unsigned c, unsigned d) const
 {
     if (kerningMapping_.empty())
         return 0;
@@ -79,7 +79,7 @@ short FontFace::GetKerning(unsigned c, unsigned d) const
 
     unsigned value = (c << 16) + d;
 
-    HashMap<unsigned, short>::const_iterator i = kerningMapping_.find(value);
+    HashMap<unsigned, float>::const_iterator i = kerningMapping_.find(value);
     if (i != kerningMapping_.end())
         return MAP_VALUE(i);
